orderManage/market/requirements.txt
psycopg[binary,pool]>=3.1.0
python-dotenv>=1.0.0
pika>=1.3.2
protobuf>=3.2
pytest>=7.0